[CLEANUP] Moved components into FSCs [WIP] Refactored Blog link [MISC] Added CSS partial for Index [CLEANUP] Changed directory structure
